+++ src/pages/Metadata/WizardComponents/Api.vue
@@ -7,21 +7,21 @@
           label="Name"
           size="large"
           element="apiName"
-          v-model="form['name']"
+          v-model="data['name']"
           description="A unique name that describes the functionality of your api in just a couple of words. Remember that this is the first description of your API that the user see and that it must convey the correct information."
         />
         <MultiLineInput
             label="Description"
             size="large"
             element="description"
-            v-model="form['description']"
+            v-model="data['description']"
             description="A short description of the functionality provided by the API"
         />
         <RadioGroup
           label="Visibility"
           name="visibility"
           :options="visibilityOptions"
-          v-model="form['visibility']"
+          v-model="data['visibility']"
           :allowCustom="false"
           customPlaceholder="Visibility"
         ></RadioGroup>
@@ -29,14 +29,14 @@
             label="System Identifier"
             size="large"
             element="systemIdentifier"
-            v-model="form['system identifier']"
+            v-model="data['system identifier']"
             description="Many APIs are developed in context of a system. This identifier is used to define this affiliation."
         />
         <InputGroup
             label="Documentation Locations"
             description="Add URLs to the API documentation. This might be OpenAPI Specifications, wiki, confluence pages etc."
             action-label="Add Additional URL"
-            v-model="form['documentation']"
+            v-model="data['documentation']"
         />
       </div>
     </div>
@@ -56,15 +56,15 @@ export default {
     RadioGroup,
     MultiLineInput
   },
-  props: { form: Object },
-  methods: {
-    updateStore() {
-      this.$store.dispatch("setEditorContent", this.form);
-    },
+  props: {
+    form: Object
   },
   computed: {
     visibilityOptions: function () {
       return this.$store.getters.metadataOptions['apiOptions'] ? this.$store.getters.metadataOptions['apiOptions']['visibilityOptions'] : [];
+    },
+    data: function(){
+      return this.form
     }
   }
 };