+++ src/utils/window-management.js
@@ -1,5 +1,8 @@
 export function scrollToPageTop() {
-    window.scroll({ top: 0, left: 0, behavior: "smooth" });
+    window.scrollTo({
+        top: 0,
+        behavior: "smooth"
+    });
 }
 
 export function scrollToComponentWithHistory(owner, refName){